Good description for body related types. Moved calls to different files for better editing
