+++ service/ldsv3/api-configurations.go
@@ -6,10 +6,37 @@ import (
 	"path"
 )
 
+// GetLogConfiguration retrieves a specific log delivery configuration.
+func (lds *Ldsv3) GetLogConfiguration(logConfigurationID string) (*ConfigurationsRespElem, error) {
+	if logConfigurationID == "" {
+		return nil, fmt.Errorf("Please provide log configuration ID")
+	}
+
+	apiURI := fmt.Sprintf("%s/log-configurations/%s", basePath, logConfigurationID)
+
+	// Create and execute request
+	resp, err := lds.Client.Rclient.R().
+		SetResult(ConfigurationsRespElem{}).
+		SetError(LsdErrorv3{}).
+		Get(apiURI)
+
+	if err != nil {
+		return nil, err
+	}
+
+	if resp.IsError() {
+		e := resp.Error().(*LsdErrorv3)
+
+		return nil, e
+	}
+
+	return resp.Result().(*ConfigurationsRespElem), nil
+}
+
 // UpdateLogConfiguration modifies a specific log delivery.
 // You need to specify all the data members in the request, or missing members are removed from the configuration.
 // The response’s Location header reflects where you can access the new configuration.
-func (lds *Ldsv3) UpdateLogConfiguration(logConfigurationID string, opts ConfigurationOptions) (string, error) {
+func (lds *Ldsv3) UpdateLogConfiguration(logConfigurationID string, body ConfigurationBody) (string, error) {
 	if logConfigurationID == "" {
 		return "", fmt.Errorf("Please provide log configuration ID")
 	}
@@ -21,7 +48,7 @@ func (lds *Ldsv3) UpdateLogConfiguration(logConfigurationID string, opts Configu
 		//SetResult(ConfigurationsRespElem{}).
 		SetError(LsdErrorv3{}).
 		SetHeader("Content-Type", "application/json").
-		SetBody(opts).
+		SetBody(body).
 		Put(apiURI)
 
 	fmt.Println(resp)
@@ -71,7 +98,7 @@ func (lds *Ldsv3) RemoveLogConfiguration(logConfigurationID string) error {
 }
 
 // CopyLogConfiguration copies a specific log delivery configuration to a target log source to produce a new log delivery configuration.
-func (lds *Ldsv3) CopyLogConfiguration(logConfigurationID string, opts ConfigurationCopyOptions) (string, error) {
+func (lds *Ldsv3) CopyLogConfiguration(logConfigurationID string, body ConfigurationCopyBody) (string, error) {
 	if logConfigurationID == "" {
 		return "", fmt.Errorf("Please provide log configuration ID")
 	}
@@ -83,7 +110,7 @@ func (lds *Ldsv3) CopyLogConfiguration(logConfigurationID string, opts Configura
 		//SetResult(ConfigurationsRespElem{}).
 		SetError(LsdErrorv3{}).
 		SetHeader("Content-Type", "application/json").
-		SetBody(opts).
+		SetBody(body).
 		Post(apiURI)
 
 	if err != nil {
@@ -164,7 +191,7 @@ func (lds *Ldsv3) ResumeLogConfiguration(logConfigurationID string) error {
 
 // CreateLogConfiguration creates new log configuration.
 // The response’s Location header reflects where you can access the new configuration.
-func (lds *Ldsv3) CreateLogConfiguration(logCSourceID, logSourceType string, opts ConfigurationOptions) (string, error) {
+func (lds *Ldsv3) CreateLogConfiguration(logCSourceID, logSourceType string, body ConfigurationBody) (string, error) {
 	if logCSourceID == "" {
 		return "", fmt.Errorf("Please provide log source ID")
 	}
@@ -173,21 +200,13 @@ func (lds *Ldsv3) CreateLogConfiguration(logCSourceID, logSourceType string, opt
 		return "", fmt.Errorf("Please provide log source type")
 	}
 
-	if logCSourceID != opts.LogSource.ID {
-		return "", fmt.Errorf("Mismatch between log source id in parameter(%s) and in request body(%s)", logCSourceID, opts.LogSource.ID)
-	}
-
-	if logSourceType != opts.LogSource.Type {
-		return "", fmt.Errorf("Mismatch between log source type in parameter(%s) and in request body(%s)", logSourceType, opts.LogSource.Type)
-	}
-
 	apiURI := fmt.Sprintf("%s/log-sources/%s/%s/log-configurations", basePath, logSourceType, logCSourceID)
 
 	// Create and execute request
 	resp, err := lds.Client.Rclient.R().
 		SetError(LsdErrorv3{}).
 		SetHeader("Content-Type", "application/json").
-		SetBody(opts).
+		SetBody(body).
 		Post(apiURI)
 
 	fmt.Println(resp)
