+++ service/ldsv3/types.go
@@ -106,55 +106,118 @@ type LogRedeliveryElem struct {
 	} `json:"links"`
 }
 
-// LogConfigurationOptions
-type ConfigurationOptions struct {
-	StartDate          string                          `json:"startDate"`
-	LogSource          ConfigurationMember             `json:"logSource"`
-	ContactDetails     ConfigurationContactDetails     `json:"contactDetails"`
-	LogFormatDetails   ConfigurationLogFormatDetails   `json:"logFormatDetails"`
-	MessageSize        ConfigurationMember             `json:"messageSize"`
-	AggregationDetails ConfigurationAggregationDetails `json:"aggregationDetails"`
-	EncodingDetails    ConfigurationEncodingDetails    `json:"encodingDetails"`
-	DeliveryDetails    ConfigurationDeliveryDetails    `json:"deliveryDetails"`
+// ConfigurationBody represents the main log delivery configuration object
+// that creates and updates a log delivery configuration.
+type ConfigurationBody struct {
+	// Start date from which logs will be collected.
+	StartDate string `json:"startDate"`
+	// (Optional) End date to which logs will be collected.
+	EndDate string `json:"endDate,omitempty"`
+	//LogSource GenericBodyMember `json:"logSource"`
+	// Contains details about contact person for this log delivery configuration.
+	ContactDetails ConfigurationBodyContactDetails `json:"contactDetails"`
+	// Describes the log format.
+	LogFormatDetails ConfigurationBodyLogFormatDetails `json:"logFormatDetails"`
+	// Packed log message’s approximate size
+	MessageSize GenericBodyMember `json:"messageSize"`
+	// Defines how to aggregate logs: by log arrival or by hit time.
+	AggregationDetails ConfigurationBodyAggregationDetails `json:"aggregationDetails"`
+	// Describes the log encoding.
+	EncodingDetails ConfigurationBodyEncodingDetails `json:"encodingDetails"`
+	// Either an email or ftp or netstorage object.
+	DeliveryDetails ConfigurationBodyDeliveryDetails `json:"deliveryDetails"`
 }
 
-// Generic Configuration type
-type ConfigurationMember struct {
-	Type string `json:"type,omitempty"`
-	ID   string `json:"id,omitempty"`
+// GenericBodyMember represents structure widely used in request body
+type GenericBodyMember struct {
+	// Unique identifier for the object
+	ID string `json:"id"`
 }
 
-type ConfigurationContactDetails struct {
-	Contact       ConfigurationMember `json:"contact"`
-	MailAddresses []string            `json:"mailAddresses"`
+// ConfigurationBodyContactDetails contains details about contact person for this log delivery configuration.
+type ConfigurationBodyContactDetails struct {
+	// Contact information provided as a simple GenericBodyMember object
+	Contact GenericBodyMember `json:"contact"`
+	// List of email addresses for contacts for this log format configuration.
+	MailAddresses []string `json:"mailAddresses"`
 }
 
-type ConfigurationLogFormatDetails struct {
-	LogFormat     ConfigurationMember `json:"logFormat"`
-	LogIdentifier string              `json:"logIdentifier"`
+// ConfigurationBodyLogFormatDetails describes the log format.
+type ConfigurationBodyLogFormatDetails struct {
+	// Selected format for log delivery, a simple GenericBodyMember object.
+	LogFormat GenericBodyMember `json:"logFormat"`
+	// Represents the first token of the log filename.
+	LogIdentifier string `json:"logIdentifier"`
 }
 
-type ConfigurationAggregationDetails struct {
-	Type              string              `json:"type"`
-	DeliveryFrequency ConfigurationMember `json:"deliveryFrequency"`
+// ConfigurationBodyAggregationDetails defines how to aggregate logs: by log arrival or by hit time.
+type ConfigurationBodyAggregationDetails struct {
+	// Identifies the type of aggregation. Possible type are byLogArrival or byHitTime
+	Type string `json:"type"`
+	// Used with byLogArrival type
+	// Period of time that will be covered by log delivery, provided as a simple GenericBodyMember object.
+	DeliveryFrequency GenericBodyMember `json:"deliveryFrequency,omitempty"`
+	// Used with byHitTime type
+	// Indicates whether residual data should be sent at regular intervals after each day.
+	DeliverResidualData bool `json:"deliverResidualData,omitempty"`
+	// Used with byHitTime type
+	// Data completion threshold, or the percentage of expected logs to be processed
+	// before the log data is sent to you, provided as a simple GenericBodyMember object.
+	DeliveryThreshold GenericBodyMember `json:"deliveryThreshold,omitempty"`
 }
 
-type ConfigurationEncodingDetails struct {
-	Encoding ConfigurationMember `json:"encoding"`
+// ConfigurationBodyEncodingDetails describes the log encoding.
+type ConfigurationBodyEncodingDetails struct {
+	// Selected encoding option used to encode logs, a simple GenericBodyMember object.
+	Encoding GenericBodyMember `json:"encoding"`
+	// Public key value for encrypted encoding.
+	// You need to set the public key value if GPG encrypted encoding is used.
+	EncodingKey string `json:"encoding,omitempty"`
 }
 
-type ConfigurationDeliveryDetails struct {
-	Type         string `json:"type"`
+// ConfigurationBodyDeliveryDetails encapsulates log delivery sent by email, ftp or netstorage(httpsns4)
+type ConfigurationBodyDeliveryDetails struct {
+	// Identifies this type of delivery, Can be email, ftp, httpsns4
+	Type string `json:"type"`
+	// Used with email type. Email address to which log will be sent.
 	EmailAddress string `json:"emailAddress,omitempty"`
+	// Used with httpsns4. Prefix of the storage group, this is part of URLs.
 	DomainPrefix string `json:"domainPrefix,omitempty"`
-	CpcodeID     int    `json:"cpcodeId,omitempty"`
-	Directory    string `json:"directory,omitempty"`
+	// Used with httpsns4. Identifies CP code within this storage group.
+	CpcodeID int `json:"cpcodeId,omitempty"`
+	// Used with httpsns4 or ftp. Directory within CP code that logs will be uploaded to.
+	Directory string `json:"directory,omitempty"`
+	// Used with ftp. Login used to authenticate to FTP machine.
+	Login string `json:"login,omitempty"`
+	// Used with ftp. Machine to which log will be sent by FTP.
+	Machine string `json:"machine,omitempty"`
+	// Used with ftp. Password used to authenticate to FTP machine.
+	// Keep in mind that this field will be empty in all responses from the server.
+	Password string `json:"password,omitempty"`
 }
 
-type ConfigurationCopyOptions struct {
-	CopyTarget ConfigurationCopyTarget `json:"copyTarget"`
+// ConfigurationCopyBody encapsulates information needed to copy a log configuration
+type ConfigurationCopyBody struct {
+	// Represents target log source for configuration copy request.
+	CopyTarget ConfigurationCopyBodyTarget `json:"copyTarget"`
 }
 
-type ConfigurationCopyTarget struct {
-	LogSource ConfigurationMember `json:"logSource"`
+// ConfigurationCopyBodyTarget represents target log source for configuration copy request.
+type ConfigurationCopyBodyTarget struct {
+	// Describes detailed log source information for configuration
+	// Both type and ID of log source are required
+	LogSource GenericBodyMember `json:"logSource"`
+}
+
+// RedeliveryBody collects information needed to create a log redelivery request.
+type RedeliveryBody struct {
+	// Log delivery configuration for which this redelivery request was created.
+	// ID in GenericBodyMember should be unique ID for the log delivery configuration.
+	LogConfiguration GenericBodyMember `json:"logConfiguration"`
+	// First hour of time range (0–23) for which log redelivery is requested.
+	BeginTime int `json:"beginTime"`
+	// Last hour of time range (1–24) for which log redelivery is requested.
+	EndTime int `json:"endTime"`
+	// Date from which log redelivery is requested.
+	RedeliveryDate string `json:"redeliveryDate"`
 }
